requirements.txt
flask==3.1.0
gunicorn==21.2.0
tensorflow==2.18.0
numpy==2.0.2
pillow==11.0.0
h5py==3.12.1
Werkzeug==3.1.3